App Class
App class resolves our issue with swapping panels. Right now its not fully implemented but its tested and seems to be able to handle multiple menus appropriately.

We are not sure on how its work with the game will be appropriate with Model view separation, but, we have limited time so that shouldn't be the main concern. Long as it works!